+++ preact.js
@@ -143,9 +143,7 @@
     return _assertThisInitialized(self);
   }
 
-  var AsyncRenderingInterrupted =
-  /*#__PURE__*/
-  function (_Error) {
+  var AsyncRenderingInterrupted = /*#__PURE__*/function (_Error) {
     _inherits(AsyncRenderingInterrupted, _Error);
 
     function AsyncRenderingInterrupted() {
@@ -159,7 +157,7 @@
     }
 
     return AsyncRenderingInterrupted;
-  }(_wrapNativeSuper(Error));
+  }( /*#__PURE__*/_wrapNativeSuper(Error));
 
   var delayWhenEmpty = 50;
   var delayWhenRendered = Infinity;
@@ -170,9 +168,7 @@
     console.error(err);
   };
 
-  var AsyncRenderingCycle =
-  /*#__PURE__*/
-  function () {
+  var AsyncRenderingCycle = /*#__PURE__*/function () {
     function AsyncRenderingCycle(target, prev, options) {
       _classCallCheck(this, AsyncRenderingCycle);
 
@@ -949,9 +945,7 @@
 
   var Component = Preact.Component;
 
-  var AsyncComponent =
-  /*#__PURE__*/
-  function (_Component) {
+  var AsyncComponent = /*#__PURE__*/function (_Component) {
     _inherits(AsyncComponent, _Component);
 
     function AsyncComponent(props) {