+++ react.mjs
@@ -179,9 +179,7 @@ function _nonIterableRest() {
   throw new TypeError("Invalid attempt to destructure non-iterable instance");
 }
 
-var AsyncRenderingInterrupted =
-/*#__PURE__*/
-function (_Error) {
+var AsyncRenderingInterrupted = /*#__PURE__*/function (_Error) {
   _inherits(AsyncRenderingInterrupted, _Error);
 
   function AsyncRenderingInterrupted() {
@@ -195,7 +193,7 @@ function (_Error) {
   }
 
   return AsyncRenderingInterrupted;
-}(_wrapNativeSuper(Error));
+}( /*#__PURE__*/_wrapNativeSuper(Error));
 
 var delayWhenEmpty = 50;
 var delayWhenRendered = Infinity;
@@ -206,9 +204,7 @@ var errorHandler = function errorHandler(err) {
   console.error(err);
 };
 
-var AsyncRenderingCycle =
-/*#__PURE__*/
-function () {
+var AsyncRenderingCycle = /*#__PURE__*/function () {
   function AsyncRenderingCycle(target, prev, options) {
     _classCallCheck(this, AsyncRenderingCycle);
 
@@ -950,7 +946,15 @@ var useState = React.useState,
     useEffect = React.useEffect;
 
 function use(asyncFunc) {
-  // create synchronous function wrapper
+  if (typeof asyncFunc !== 'function') {
+    if (process.env.NODE_ENV !== 'production') {
+      console.warn('Non-function given to Relaks.use():', asyncFunc);
+    }
+
+    return asyncFunc;
+  } // create synchronous function wrapper
+
+
   var syncFunc = function syncFunc(props, ref) {
     var state = useState({});
     var target = {
@@ -1032,7 +1036,13 @@ function use(asyncFunc) {
   } // set display name
 
 
-  syncFunc.displayName = asyncFunc.displayName || asyncFunc.name;
+  if (asyncFunc.name) {
+    Object.defineProperty(syncFunc, 'name', {
+      value: asyncFunc.name,
+      writable: false
+    });
+  }
+
   return syncFunc;
 }
 
@@ -1102,9 +1112,7 @@ var functions = /*#__PURE__*/Object.freeze({
 
 var PureComponent = React.PureComponent;
 
-var AsyncComponent =
-/*#__PURE__*/
-function (_PureComponent) {
+var AsyncComponent = /*#__PURE__*/function (_PureComponent) {
   _inherits(AsyncComponent, _PureComponent);
 
   function AsyncComponent(props) {
@@ -1233,9 +1241,7 @@ function useRenderEvent(name, f) {
   }
 }
 
-var AsyncSaveBuffer =
-/*#__PURE__*/
-function () {
+var AsyncSaveBuffer = /*#__PURE__*/function () {
   function AsyncSaveBuffer() {
     _classCallCheck(this, AsyncSaveBuffer);